# misc/swagger.yaml
swagger: "2.0"
info:
  description: "This is CDTPortal you can create projects for all tools at once."
  version: "1.0.0"
  title: "CDTPortal"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "admin@devopswise.co.uk"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "cdtportal.cdt.devopswise.co.uk"
basePath: "/v2"
...
schemes:
- "https"
- "http"
paths:
...
  /workspace:
    post:
      tags:
        - workspace
      summary: create a new workspace
      description: 'Creates a workspace shared between all tools'
      operationId: createWorkspace
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Workspace description to create
          required: true
          schema:
            $ref: '#/definitions/Workspace'
      responses:
        '405':
          description: Invalid input
      security:
        - api_key: []
        - cdtportal_auth: []
    put:
      tags:
        - workspace
      summary: Update an existing workspace
      description: ''
      operationId: updateWorkspace
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: workspace object that needs to be added to the store
          required: true
          schema:
            $ref: '#/definitions/Workspace'
      responses:
        '400':
          description: Invalid ID supplied
        '404':
          description: workspace not found
        '405':
          description: Validation exception
      security:
        - api_key: []
        - cdtportal_auth: []
  /workspace/findByStatus:
    get:
      tags:
        - workspace
      summary: Finds workspaces by status
      description: Multiple status values can be provided with comma separated strings
      operationId: findWorkspacesByStatus
      produces:
        - application/json
      parameters:
        - name: status
          in: query
          description: Status values that need to be considered for filter
          required: true
          type: array
          items:
            type: string
            enum:
              - available
              - pending
              - sold
            default: available
          collectionFormat: csv
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Workspace'
        '400':
          description: Invalid status value
      security:
        - api_key: []
        - cdtportal_auth: []
  /workspace/findByOwner:
    get:
      tags:
        - workspace
      summary: Finds workspaces by its owner
      description: List workspaces of given user
      operationId: findWorkspacesByOwner
      produces:
        - application/json
      parameters:
        - name: username
          in: query
          description: username of workspace owner
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Workspace'
        '400':
          description: Invalid status value
      security:
        - api_key: []
        - cdtportal_auth: []
  /workspace/findByName:
    get:
      tags:
        - workspace
      summary: Finds workspaces by its name
      description: List workspaces meeting criteria
      operationId: findWorkspacesByName
      produces:
        - application/json
      parameters:
        - name: workspaceName
          in: query
          description: name of the workspace
          required: true
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Workspace'
        '400':
          description: Invalid status value
      security:
        - api_key: []
        - cdtportal_auth: []
  /workspace/{workspaceId}:
    get:
      tags:
        - workspace
      summary: Find workspace by ID
      description: Returns a single workspace
      operationId: getWorkspaceById
      produces:
        - application/json
      parameters:
        - name: workspaceId
          in: path
          description: ID of workspace to return
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Workspace'
        '400':
          description: Invalid ID supplied
        '404':
          description: workspace not found
      security:
        - api_key: []
        - cdtportal_auth: []
    post:
      tags:
        - workspace
      summary: Updates a workspace in the store with form data
      description: ''
      operationId: updateWorkspaceWithForm
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: workspaceId
          in: path
          description: ID of workspace that needs to be updated
          required: true
          type: integer
          format: int64
        - name: name
          in: formData
          description: Updated name of the workspace
          required: false
          type: string
        - name: status
          in: formData
          description: Updated status of the workspace
          required: false
          type: string
      responses:
        '405':
          description: Invalid input
      security:
        - api_key: []
        - cdtportal_auth: []
    delete:
      tags:
        - workspace
      summary: Deletes a workspace
      description: ''
      operationId: deleteWorkspace
      produces:
        - application/json
      parameters:
        - name: api_key
          in: header
          required: false
          type: string
        - name: workspaceId
          in: path
          description: workspace id to delete
          required: true
          type: integer
          format: int64
      responses:
        '400':
          description: Invalid workspace value
      security:
        - api_key: []
        - cdtportal_auth: []
...
securityDefinitions:
  cdtportal_auth:
    type: "basic"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Tag:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
...
  Workspace:
    title: Workspace
    description: A workspace based on Theia ide
    type: object
    required:
      - name
      - key
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
        example: Helloworld
      owner:
        type: string
      git_url:
        type: string
        example: git://hrweb@gitea
      tags:
        type: array
        items:
          $ref: '#/definitions/Tag'
      description:
        type: string
        description: definition of project, like Human Resources Web Portal
...